refactor: renamed game_engine
